+++ ed.py
@@ -17,10 +17,10 @@ def transverse_ising_hamiltonian(N: int, J: float, h: float):
     if N <= 0:
         raise ValueError("N must be positive")
 
-    H = np.zeros((2**N, 2**N), dtype=np.complex64)
-    X = np.array([[0.0, 1.0], [1.0, 0.0]], dtype=np.complex64)
-    Z = np.array([[1.0, 0.0], [0.0, -1.0]], dtype=np.complex64)
-    I = np.array([[1.0, 0.0], [0.0, 1.0]], dtype=np.complex64)
+    H = np.zeros((2**N, 2**N), dtype=np.complex128)
+    X = np.array([[0.0, 1.0], [1.0, 0.0]], dtype=np.complex128)
+    Z = np.array([[1.0, 0.0], [0.0, -1.0]], dtype=np.complex128)
+    I = np.array([[1.0, 0.0], [0.0, 1.0]], dtype=np.complex128)
 
     for i in range(N):
         # Sz-Sz term with periodic boundary conditions
@@ -39,7 +39,7 @@ def transverse_ising_hamiltonian(N: int, J: float, h: float):
 
 def thermal_state(H: np.ndarray, beta: float):
     """Return the Gibbs state and thermal energy for a Hermitian matrix."""
-    H = np.asarray(H, dtype=np.complex64)
+    H = np.asarray(H, dtype=np.complex128)
     if H.shape[0] != H.shape[1]:
         raise ValueError("H must be square")
 
@@ -47,16 +47,16 @@ def thermal_state(H: np.ndarray, beta: float):
     shifted_evals = evals - np.min(evals)
     beta_exps = np.exp(-beta * shifted_evals)
     Z = np.sum(beta_exps)
-    weights = np.asarray(beta_exps / Z, dtype=np.complex64)
-    gibbs = ((evecs.astype(np.complex64) * weights) @ evecs.conj().T).astype(np.complex64)
-    energy = np.complex64(np.sum(evals * beta_exps) / Z)
+    weights = np.asarray(beta_exps / Z, dtype=np.complex128)
+    gibbs = ((evecs.astype(np.complex128) * weights) @ evecs.conj().T).astype(np.complex128)
+    energy = np.complex128(np.sum(evals * beta_exps) / Z)
     return gibbs, energy
 
 
 def thermal_expectation_value(H: np.ndarray, A: np.ndarray, beta: float):
     """Return the thermal expectation value of A."""
-    H = np.asarray(H, dtype=np.complex64)
-    A = np.asarray(A, dtype=np.complex64)
+    H = np.asarray(H, dtype=np.complex128)
+    A = np.asarray(A, dtype=np.complex128)
     if H.shape[0] != H.shape[1]:
         raise ValueError("H must be square")
 
@@ -65,33 +65,33 @@ def thermal_expectation_value(H: np.ndarray, A: np.ndarray, beta: float):
     beta_exps = np.exp(-beta * shifted_evals)
     Z = np.sum(beta_exps)
 
-    expectation = np.complex64(0.0)
+    expectation = np.complex128(0.0)
     for i in range(H.shape[0]):
         ket = evecs[:, i]
-        expectation += np.complex64(beta_exps[i] / Z) * np.complex64(ket.conj().T @ A @ ket)
-    return np.complex64(expectation)
+        expectation += np.complex128(beta_exps[i] / Z) * np.complex128(ket.conj().T @ A @ ket)
+    return np.complex128(expectation)
 
 
 def ground_state(H: np.ndarray):
     """Return the ground-state vector and energy."""
-    H = np.asarray(H, dtype=np.complex64)
+    H = np.asarray(H, dtype=np.complex128)
     if H.shape[0] != H.shape[1]:
         raise ValueError("H must be square")
 
     evals, evecs = np.linalg.eigh(H)
-    return evecs[:, 0].astype(np.complex64), np.complex64(evals[0])
+    return evecs[:, 0].astype(np.complex128), np.complex128(evals[0])
 
 
 def ground_state_expectation_value(H: np.ndarray, A: np.ndarray):
     """Return the ground-state expectation value of A."""
-    H = np.asarray(H, dtype=np.complex64)
-    A = np.asarray(A, dtype=np.complex64)
+    H = np.asarray(H, dtype=np.complex128)
+    A = np.asarray(A, dtype=np.complex128)
     if H.shape[0] != H.shape[1]:
         raise ValueError("H must be square")
 
     _, evecs = np.linalg.eigh(H)
-    gs = evecs[:, 0].astype(np.complex64)
-    return np.complex64(gs.conj().T @ A @ gs)
+    gs = evecs[:, 0].astype(np.complex128)
+    return np.complex128(gs.conj().T @ A @ gs)
 
 
 def get_transverse_ising_groundstate(N: int, J: float, h: float):
@@ -108,7 +108,7 @@ def get_transverse_ising_gibbsstate(N: int, J: float, h: float, beta: float):
 
 def get_spectrum(H: np.ndarray):
     """Return the eigenvalues of H in ascending order."""
-    H = np.asarray(H, dtype=np.complex64)
+    H = np.asarray(H, dtype=np.complex128)
     if H.shape[0] != H.shape[1]:
         raise ValueError("H must be square")
     return np.linalg.eigvalsh(H)
